Fix delete card

+++ controllers/cards.js
@@ -23,9 +23,18 @@ const getCards = (req,res) => {
 
 const deleteCardById = (req, res) => {
   const {id} = req.params;
-  console.log(id);
+  if (!validateId(id)) {
+    return res.status(400).send({
+      "message": "Передан некорректный ID"
+    })
+  }
   return Card.deleteOne({_id: id})
     .then((card) => {
+      if (!card) {
+        return res.status(404).send({
+          "message": "Карточка не найдена"
+        })
+      }
       return res.status(200).send(card);
     })
     .catch((err) => {
